+++ mobile/src/androidTest/java/treehou/se/habit/data/TestConnectionFactory.java
@@ -4,7 +4,9 @@ import android.content.Context;
 
 import java.util.List;
 
+import okhttp3.ResponseBody;
 import retrofit2.Call;
+import retrofit2.Response;
 import rx.Observable;
 import se.treehou.ng.ohcommunicator.connector.models.OHBinding;
 import se.treehou.ng.ohcommunicator.connector.models.OHInboxItem;
@@ -13,6 +15,7 @@ import se.treehou.ng.ohcommunicator.connector.models.OHLink;
 import se.treehou.ng.ohcommunicator.connector.models.OHLinkedPage;
 import se.treehou.ng.ohcommunicator.connector.models.OHServer;
 import se.treehou.ng.ohcommunicator.connector.models.OHSitemap;
+import se.treehou.ng.ohcommunicator.connector.models.OHThing;
 import se.treehou.ng.ohcommunicator.services.Connector;
 import se.treehou.ng.ohcommunicator.services.IServerHandler;
 import se.treehou.ng.ohcommunicator.services.callbacks.OHCallback;
@@ -52,12 +55,27 @@ public class TestConnectionFactory extends ConnectionFactory {
         }
 
         @Override
-        public Call<Void> createLink(OHLink ohLink) {
+        public void createLink(OHLink ohLink) {
+
+        }
+
+        @Override
+        public Observable<Response<ResponseBody>> createLinkRx(OHLink ohLink) {
             return null;
         }
 
         @Override
-        public Call<Void> deleteLink(OHLink ohLink) {
+        public void deleteLink(OHLink ohLink) {
+
+        }
+
+        @Override
+        public Observable<Response<ResponseBody>> deleteLinkRx(OHLink ohLink) {
+            return null;
+        }
+
+        @Override
+        public Observable<List<OHThing>> requestThingsRx() {
             return null;
         }